+++ NEWS.md
@@ -1,3 +1,7 @@
+# fireproof (development version)
+
+* Add support for `guard_auth0()`
+
 # fireproof 0.1.0
 
 * Initial CRAN submission.